requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pyyaml
numpy
